pyproject.toml
[tool.poetry]
name = "coure-work-oop"
version = "0.1.0"
description = ""
authors = ["Roman Gerasimov <rgerasimov58@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.31.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
